Move itemdto to base models folder
