+++ src/components/cardbuilder/Card/Card.tsx
@@ -4,7 +4,7 @@ import CardWrapper from './CardWrapper';
 import CardBox from './CardBox';
 
 import type { CardOptions } from 'types/cardOptions';
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 
 interface CardProps {
     item?: ItemDto;
