+++ src/components/cardbuilder/Card/CardContent.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import { getDefaultBackgroundClass } from '../cardBuilderUtils';
 import CardImageContainer from './CardImageContainer';
 
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 import type { CardOptions } from 'types/cardOptions';
 
 interface CardContentProps {
