+++ src/components/cardbuilder/Card/CardInnerFooter.tsx
@@ -1,7 +1,7 @@
 import React, { type FC } from 'react';
 import classNames from 'classnames';
 import CardFooterText from './CardFooterText';
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 import type { CardOptions } from 'types/cardOptions';
 
 interface CardInnerFooterProps {
