+++ src/components/cardbuilder/Card/CardOverlayButtons.tsx
@@ -7,7 +7,7 @@ import { appRouter } from 'components/router/appRouter';
 import PlayArrowIconButton from '../../common/PlayArrowIconButton';
 import MoreVertIconButton from '../../common/MoreVertIconButton';
 
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 import type { CardOptions } from 'types/cardOptions';
 
 const sholudShowOverlayPlayButton = (
