+++ src/components/cardbuilder/Card/useCardText.tsx
@@ -5,7 +5,7 @@ import layoutManager from 'components/layoutManager';
 import CardText from './CardText';
 import { getCardTextLines } from './cardHelper';
 
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 import type { CardOptions } from 'types/cardOptions';
 
 const enableRightMargin = (
