+++ src/components/common/DefaultName.tsx
@@ -2,7 +2,7 @@ import React, { type FC } from 'react';
 import Box from '@mui/material/Box';
 import itemHelper from 'components/itemHelper';
 import { isUsingLiveTvNaming } from '../cardbuilder/cardBuilderUtils';
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 
 interface DefaultNameProps {
     item: ItemDto;
