+++ src/components/indicators/useIndicator.tsx
@@ -17,7 +17,7 @@ import datetime from 'scripts/datetime';
 import itemHelper from 'components/itemHelper';
 import AutoTimeProgressBar from 'elements/emby-progressbar/AutoTimeProgressBar';
 import type { NullableString } from 'types/base/common/shared/types';
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 import type { ProgressOptions } from 'types/progressOptions';
 
 const TypeIcon = {
