+++ src/components/listview/List/useList.ts
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import { getDataAttributes } from 'utils/items';
 import layoutManager from 'components/layoutManager';
 
-import type { ItemDto } from 'types/itemDto';
+import type { ItemDto } from 'types/base/models/item-dto';
 import type { ListOptions } from 'types/listOptions';
 
 interface UseListProps {
