+++ src/types/base/models/item-dto.ts
