+++ src/types/listOptions.ts
@@ -1,6 +1,6 @@
 import { ItemSortBy } from '@jellyfin/sdk/lib/models/api/item-sort-by';
 import type { CollectionType } from '@jellyfin/sdk/lib/generated-client/models/collection-type';
-import type { ItemDto } from './itemDto';
+import type { ItemDto } from './base/models/item-dto';
 export interface ListOptions {
     items?: ItemDto[] | null;
     index?: string;
